Refactored containerHandler()
`containerHandler()` was much too large. This change moves all of the
underlying User Container code into it's own helper file. To reduce the
number of function parameters (e.g. workflow, profile, ctx) passed into
these functions, a new struct was added so that the helper functions
could make use it.

Also renamed it to `userContainerHandler()`.
